android-game-sdk-rs: bump v2.0.2 -> v4.0.0
