+++ android-activity/android-games-sdk/file_list.txt
@@ -1,11 +1,13 @@
-GameActivity.h
+android_native_app_glue.c
+android_native_app_glue.h
 GameActivity.cpp
-GameActivityEvents.h
 GameActivityEvents.cpp
+GameActivityEvents.h
+GameActivityEvents_internal.h
+GameActivity.h
 GameActivityLog.h
+gamesdk_common.h
 gametextinput.cpp
 gametextinput.h
-gamecommon.h
-gamesdk_common.h
-android_native_app_glue.h
-android_native_app_glue.c
+system_utils.cpp
+system_utils.h
