+++ android-activity/android-games-sdk/game-activity/prefab-src/modules/game-activity/include/game-activity/GameActivityEvents.h
@@ -57,29 +57,29 @@ extern "C" {
  * \see GameActivityMotionEvent
  */
 typedef struct GameActivityPointerAxes {
-    int32_t id;
-    int32_t toolType;
-    float axisValues[GAME_ACTIVITY_POINTER_INFO_AXIS_COUNT];
-    float rawX;
-    float rawY;
+  int32_t id;
+  int32_t toolType;
+  float axisValues[GAME_ACTIVITY_POINTER_INFO_AXIS_COUNT];
+  float rawX;
+  float rawY;
 } GameActivityPointerAxes;
 
 /** \brief Get the toolType of the pointer. */
 inline int32_t GameActivityPointerAxes_getToolType(
     const GameActivityPointerAxes* pointerInfo) {
-    return pointerInfo->toolType;
+  return pointerInfo->toolType;
 }
 
 /** \brief Get the current X coordinate of the pointer. */
 inline float GameActivityPointerAxes_getX(
     const GameActivityPointerAxes* pointerInfo) {
-    return pointerInfo->axisValues[AMOTION_EVENT_AXIS_X];
+  return pointerInfo->axisValues[AMOTION_EVENT_AXIS_X];
 }
 
 /** \brief Get the current Y coordinate of the pointer. */
 inline float GameActivityPointerAxes_getY(
     const GameActivityPointerAxes* pointerInfo) {
-    return pointerInfo->axisValues[AMOTION_EVENT_AXIS_Y];
+  return pointerInfo->axisValues[AMOTION_EVENT_AXIS_Y];
 }
 
 /**
@@ -126,44 +126,44 @@ float GameActivityPointerAxes_getAxisValue(
 
 inline float GameActivityPointerAxes_getPressure(
     const GameActivityPointerAxes* pointerInfo) {
-    return GameActivityPointerAxes_getAxisValue(pointerInfo,
-                                                AMOTION_EVENT_AXIS_PRESSURE);
+  return GameActivityPointerAxes_getAxisValue(pointerInfo,
+                                              AMOTION_EVENT_AXIS_PRESSURE);
 }
 
 inline float GameActivityPointerAxes_getSize(
     const GameActivityPointerAxes* pointerInfo) {
-    return GameActivityPointerAxes_getAxisValue(pointerInfo,
-                                                AMOTION_EVENT_AXIS_SIZE);
+  return GameActivityPointerAxes_getAxisValue(pointerInfo,
+                                              AMOTION_EVENT_AXIS_SIZE);
 }
 
 inline float GameActivityPointerAxes_getTouchMajor(
     const GameActivityPointerAxes* pointerInfo) {
-    return GameActivityPointerAxes_getAxisValue(pointerInfo,
-                                                AMOTION_EVENT_AXIS_TOUCH_MAJOR);
+  return GameActivityPointerAxes_getAxisValue(pointerInfo,
+                                              AMOTION_EVENT_AXIS_TOUCH_MAJOR);
 }
 
 inline float GameActivityPointerAxes_getTouchMinor(
     const GameActivityPointerAxes* pointerInfo) {
-    return GameActivityPointerAxes_getAxisValue(pointerInfo,
-                                                AMOTION_EVENT_AXIS_TOUCH_MINOR);
+  return GameActivityPointerAxes_getAxisValue(pointerInfo,
+                                              AMOTION_EVENT_AXIS_TOUCH_MINOR);
 }
 
 inline float GameActivityPointerAxes_getToolMajor(
     const GameActivityPointerAxes* pointerInfo) {
-    return GameActivityPointerAxes_getAxisValue(pointerInfo,
-                                                AMOTION_EVENT_AXIS_TOOL_MAJOR);
+  return GameActivityPointerAxes_getAxisValue(pointerInfo,
+                                              AMOTION_EVENT_AXIS_TOOL_MAJOR);
 }
 
 inline float GameActivityPointerAxes_getToolMinor(
     const GameActivityPointerAxes* pointerInfo) {
-    return GameActivityPointerAxes_getAxisValue(pointerInfo,
-                                                AMOTION_EVENT_AXIS_TOOL_MINOR);
+  return GameActivityPointerAxes_getAxisValue(pointerInfo,
+                                              AMOTION_EVENT_AXIS_TOOL_MINOR);
 }
 
 inline float GameActivityPointerAxes_getOrientation(
     const GameActivityPointerAxes* pointerInfo) {
-    return GameActivityPointerAxes_getAxisValue(pointerInfo,
-                                                AMOTION_EVENT_AXIS_ORIENTATION);
+  return GameActivityPointerAxes_getAxisValue(pointerInfo,
+                                              AMOTION_EVENT_AXIS_ORIENTATION);
 }
 
 /**
@@ -171,7 +171,7 @@ inline float GameActivityPointerAxes_getOrientation(
  */
 #if (defined GAMEACTIVITY_MAX_NUM_POINTERS_IN_MOTION_EVENT_OVERRIDE)
 #define GAMEACTIVITY_MAX_NUM_POINTERS_IN_MOTION_EVENT \
-    GAMEACTIVITY_MAX_NUM_POINTERS_IN_MOTION_EVENT_OVERRIDE
+  GAMEACTIVITY_MAX_NUM_POINTERS_IN_MOTION_EVENT_OVERRIDE
 #else
 #define GAMEACTIVITY_MAX_NUM_POINTERS_IN_MOTION_EVENT 8
 #endif
@@ -183,32 +183,32 @@ inline float GameActivityPointerAxes_getOrientation(
  * (see https://developer.android.com/reference/android/view/MotionEvent).
  */
 typedef struct GameActivityMotionEvent {
-    int32_t deviceId;
-    int32_t source;
-    int32_t action;
+  int32_t deviceId;
+  int32_t source;
+  int32_t action;
 
-    int64_t eventTime;
-    int64_t downTime;
+  int64_t eventTime;
+  int64_t downTime;
 
-    int32_t flags;
-    int32_t metaState;
+  int32_t flags;
+  int32_t metaState;
 
-    int32_t actionButton;
-    int32_t buttonState;
-    int32_t classification;
-    int32_t edgeFlags;
+  int32_t actionButton;
+  int32_t buttonState;
+  int32_t classification;
+  int32_t edgeFlags;
 
-    uint32_t pointerCount;
-    GameActivityPointerAxes
-        pointers[GAMEACTIVITY_MAX_NUM_POINTERS_IN_MOTION_EVENT];
+  uint32_t pointerCount;
+  GameActivityPointerAxes
+      pointers[GAMEACTIVITY_MAX_NUM_POINTERS_IN_MOTION_EVENT];
 
-    int historySize;
-    int64_t* historicalEventTimesMillis;
-    int64_t* historicalEventTimesNanos;
-    float* historicalAxisValues;
+  int historySize;
+  int64_t* historicalEventTimesMillis;
+  int64_t* historicalEventTimesNanos;
+  float* historicalAxisValues;
 
-    float precisionX;
-    float precisionY;
+  float precisionX;
+  float precisionY;
 } GameActivityMotionEvent;
 
 float GameActivityMotionEvent_getHistoricalAxisValue(
@@ -217,77 +217,65 @@ float GameActivityMotionEvent_getHistoricalAxisValue(
 
 inline int GameActivityMotionEvent_getHistorySize(
     const GameActivityMotionEvent* event) {
-    return event->historySize;
+  return event->historySize;
 }
 
 inline float GameActivityMotionEvent_getHistoricalX(
     const GameActivityMotionEvent* event, int pointerIndex, int historyPos) {
-    return GameActivityMotionEvent_getHistoricalAxisValue(
-        event, AMOTION_EVENT_AXIS_X, pointerIndex, historyPos);
+  return GameActivityMotionEvent_getHistoricalAxisValue(
+      event, AMOTION_EVENT_AXIS_X, pointerIndex, historyPos);
 }
 
 inline float GameActivityMotionEvent_getHistoricalY(
     const GameActivityMotionEvent* event, int pointerIndex, int historyPos) {
-    return GameActivityMotionEvent_getHistoricalAxisValue(
-        event, AMOTION_EVENT_AXIS_Y, pointerIndex, historyPos);
+  return GameActivityMotionEvent_getHistoricalAxisValue(
+      event, AMOTION_EVENT_AXIS_Y, pointerIndex, historyPos);
 }
 
 inline float GameActivityMotionEvent_getHistoricalPressure(
     const GameActivityMotionEvent* event, int pointerIndex, int historyPos) {
-    return GameActivityMotionEvent_getHistoricalAxisValue(
-        event, AMOTION_EVENT_AXIS_PRESSURE, pointerIndex, historyPos);
+  return GameActivityMotionEvent_getHistoricalAxisValue(
+      event, AMOTION_EVENT_AXIS_PRESSURE, pointerIndex, historyPos);
 }
 
 inline float GameActivityMotionEvent_getHistoricalSize(
     const GameActivityMotionEvent* event, int pointerIndex, int historyPos) {
-    return GameActivityMotionEvent_getHistoricalAxisValue(
-        event, AMOTION_EVENT_AXIS_SIZE, pointerIndex, historyPos);
+  return GameActivityMotionEvent_getHistoricalAxisValue(
+      event, AMOTION_EVENT_AXIS_SIZE, pointerIndex, historyPos);
 }
 
 inline float GameActivityMotionEvent_getHistoricalTouchMajor(
     const GameActivityMotionEvent* event, int pointerIndex, int historyPos) {
-    return GameActivityMotionEvent_getHistoricalAxisValue(
-        event, AMOTION_EVENT_AXIS_TOUCH_MAJOR, pointerIndex, historyPos);
+  return GameActivityMotionEvent_getHistoricalAxisValue(
+      event, AMOTION_EVENT_AXIS_TOUCH_MAJOR, pointerIndex, historyPos);
 }
 
 inline float GameActivityMotionEvent_getHistoricalTouchMinor(
     const GameActivityMotionEvent* event, int pointerIndex, int historyPos) {
-    return GameActivityMotionEvent_getHistoricalAxisValue(
-        event, AMOTION_EVENT_AXIS_TOUCH_MINOR, pointerIndex, historyPos);
+  return GameActivityMotionEvent_getHistoricalAxisValue(
+      event, AMOTION_EVENT_AXIS_TOUCH_MINOR, pointerIndex, historyPos);
 }
 
 inline float GameActivityMotionEvent_getHistoricalToolMajor(
     const GameActivityMotionEvent* event, int pointerIndex, int historyPos) {
-    return GameActivityMotionEvent_getHistoricalAxisValue(
-        event, AMOTION_EVENT_AXIS_TOOL_MAJOR, pointerIndex, historyPos);
+  return GameActivityMotionEvent_getHistoricalAxisValue(
+      event, AMOTION_EVENT_AXIS_TOOL_MAJOR, pointerIndex, historyPos);
 }
 
 inline float GameActivityMotionEvent_getHistoricalToolMinor(
     const GameActivityMotionEvent* event, int pointerIndex, int historyPos) {
-    return GameActivityMotionEvent_getHistoricalAxisValue(
-        event, AMOTION_EVENT_AXIS_TOOL_MINOR, pointerIndex, historyPos);
+  return GameActivityMotionEvent_getHistoricalAxisValue(
+      event, AMOTION_EVENT_AXIS_TOOL_MINOR, pointerIndex, historyPos);
 }
 
 inline float GameActivityMotionEvent_getHistoricalOrientation(
     const GameActivityMotionEvent* event, int pointerIndex, int historyPos) {
-    return GameActivityMotionEvent_getHistoricalAxisValue(
-        event, AMOTION_EVENT_AXIS_ORIENTATION, pointerIndex, historyPos);
+  return GameActivityMotionEvent_getHistoricalAxisValue(
+      event, AMOTION_EVENT_AXIS_ORIENTATION, pointerIndex, historyPos);
 }
 
 /** \brief Handle the freeing of the GameActivityMotionEvent struct. */
 void GameActivityMotionEvent_destroy(GameActivityMotionEvent* c_event);
-
-/**
- * \brief Convert a Java `MotionEvent` to a `GameActivityMotionEvent`.
- *
- * This is done automatically by the GameActivity: see `onTouchEvent` to set
- * a callback to consume the received events.
- * This function can be used if you re-implement events handling in your own
- * activity.
- * Ownership of out_event is maintained by the caller.
- */
-void GameActivityMotionEvent_fromJava(JNIEnv* env, jobject motionEvent,
-                                      GameActivityMotionEvent* out_event);
 
 /**
  * \brief Describe a key event that happened on the GameActivity SurfaceView.
@@ -298,34 +286,22 @@ void GameActivityMotionEvent_fromJava(JNIEnv* env, jobject motionEvent,
  * nanoseconds in this struct.
  */
 typedef struct GameActivityKeyEvent {
-    int32_t deviceId;
-    int32_t source;
-    int32_t action;
+  int32_t deviceId;
+  int32_t source;
+  int32_t action;
 
-    int64_t eventTime;
-    int64_t downTime;
+  int64_t eventTime;
+  int64_t downTime;
 
-    int32_t flags;
-    int32_t metaState;
+  int32_t flags;
+  int32_t metaState;
 
-    int32_t modifiers;
-    int32_t repeatCount;
-    int32_t keyCode;
-    int32_t scanCode;
-    //int32_t unicodeChar;
+  int32_t modifiers;
+  int32_t repeatCount;
+  int32_t keyCode;
+  int32_t scanCode;
+  // int32_t unicodeChar;
 } GameActivityKeyEvent;
-
-/**
- * \brief Convert a Java `KeyEvent` to a `GameActivityKeyEvent`.
- *
- * This is done automatically by the GameActivity: see `onKeyUp` and `onKeyDown`
- * to set a callback to consume the received events.
- * This function can be used if you re-implement events handling in your own
- * activity.
- * Ownership of out_event is maintained by the caller.
- */
-void GameActivityKeyEvent_fromJava(JNIEnv* env, jobject motionEvent,
-                                   GameActivityKeyEvent* out_event);
 
 #ifdef __cplusplus
 }
